Only delete isolate box if exists

+++ judgels-backends/gabriel/gabriel-engines/src/main/java/judgels/gabriel/sandboxes/moe/IsolateSandbox.java
@@ -288,7 +288,9 @@ public class IsolateSandbox implements Sandbox {
             if (result.getExitCode() != 0) {
                 throw new SandboxException("Cannot clean up Isolate!");
             }
-            FileUtils.forceDelete(boxDir);
+            if (boxDir.exists()) {
+                FileUtils.forceDelete(boxDir);
+            }
         } catch (IOException | InterruptedException e) {
             throw new SandboxException(e);
         }
